backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.5
djangorestframework==3.14.0
beautifulsoup4
psycopg2-binary
django-environ
gunicorn==21.2.0
requests
celery==5.2.3
redis==4.0.2
django-celery-beat
django-celery-results
